// packages/arbitrum-dai-bridge/typechain-types/L1DaiTeleportGateway.ts
export type TeleportGUIDStruct = {
  sourceDomain: BytesLike;
  targetDomain: BytesLike;
  receiver: BytesLike;
  operator: BytesLike;
  amount: BigNumberish;
  nonce: BigNumberish;
  timestamp: BigNumberish;
};